--- a/mathutil/sets_test.go
+++ b/mathutil/sets_test.go
@@ -48,6 +48,13 @@ func TestComplementsInt(t *testing.T) {
 			aExpected: []int64{1, 2},
 			bExpected: []int64{5, 6},
 		},
+		{
+			name:      "Overlap with repeated values",
+			a:         []int64{6, 4, 5, 3, 6},
+			b:         []int64{2, 1, 4, 3, 1},
+			aExpected: []int64{5, 6},
+			bExpected: []int64{1, 2},
+		},
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
@@ -61,5 +68,32 @@ func TestComplementsInt(t *testing.T) {
 				t.Errorf("B differs:\n%s", d)
 			}
 		})
+	}
+}
+
+func BenchmarkComplementsInt_equal(b *testing.B) {
+	listA := []int64{1, 2, 3}
+	listB := []int64{1, 2, 3}
+
+	for n := 0; n < b.N; n++ {
+		ComplementsInt(listA, listB)
+	}
+}
+
+func BenchmarkComplementsInt_disjoint(b *testing.B) {
+	listA := []int64{1, 2, 3}
+	listB := []int64{5, 6, 7}
+
+	for n := 0; n < b.N; n++ {
+		ComplementsInt(listA, listB)
+	}
+}
+
+func BenchmarkComplementsInt_overlap(b *testing.B) {
+	listA := []int64{1, 2, 3, 4}
+	listB := []int64{3, 4, 5, 6}
+
+	for n := 0; n < b.N; n++ {
+		ComplementsInt(listA, listB)
 	}
 }