To keep the same behavior on the returned slices of the ComplementsInt
function, we should remove the repeated values from the `bOnly` slice.

We used the same approach of the `aOnly` slice, using a map to guarantee only
unique elements. The downside is a worse performance as shown below.

Before:
```
BenchmarkComplementsInt_equal-8    10000000 217 ns/op   0 B/op  0 allocs/op
BenchmarkComplementsInt_disjoint-8  3000000 484 ns/op 160 B/op  8 allocs/op
BenchmarkComplementsInt_overlap-8   3000000 498 ns/op  96 B/op  6 allocs/op
```

After:
```
BenchmarkComplementsInt_equal-8     5000000 255 ns/op   0 B/op  0 allocs/op
BenchmarkComplementsInt_disjoint-8  2000000 769 ns/op 208 B/op 10 allocs/op
BenchmarkComplementsInt_overlap-8   2000000 717 ns/op 144 B/op  8 allocs/op
```