@rafaeljusto 2017-12-28
If time performance is an issue, I worked on a [new solution](https://github.com/rafaeljusto/utils/blob/mathutil-performance/mathutil/sets.go) replacing maps for sorted slices. The code has better time performance results, but it allocates more memory:

```
BenchmarkComplementsInt_equal-8    10000000 186 ns/op 128 B/op  4 allocs/op
BenchmarkComplementsInt_disjoint-8  5000000 352 ns/op 240 B/op 10 allocs/op
BenchmarkComplementsInt_overlap-8   5000000 300 ns/op 176 B/op  8 allocs/op
```

Also, the code complexity increased, making it more difficult for some future maintenance. Let me known if you guys want a PR.