@Carpetsmoker 2017-12-28
Thanks for your PR!

I'm not entirely sure this is a good idea though. `[]int64` lists aren't "real" sets, as you can insert duplicate items and people may expect them to be in the same order. This function is just a simple helper to get a "set-like functionality".

If you want "true" sets, then it's probably best to use `map[int64]struct{}`, or [golang-set](https://github.com/deckarep/golang-set) from the start, rather than `[]int64` (...and convert this []int64 to a map and back again when doing set-like operations).

---

I created #20 which moves the function to `sliceutil.Complement`, and changes the implementation to one that also preserves the order of the lists. This is a bit slower, but that is not a huge deal for most cases.

You can use `sliceutil.UniqInt64` to remove duplicates btw.