@rafaeljusto 2017-12-28
Thanks @Carpetsmoker ! It makes sense to keep the order of the slices.